Feat : 페이지 예외처리 메서드 추가 및 클래스 위치 수정
Related:#81
Ref:#54
